create_ce_tree_long_df

Completely specifying cost-effectiveness decision tree. Matches labels on branches to the corresponding probabilities and cost/utility values.

create_ce_tree_long_df(
  tree_list = NA,
  tree_mat = NA,
  label_probs,
  label_costs,
  label_health,
  pname_from_to,
  cname_from_to,
  hname_from_to
)

Arguments

tree_list
Parent-child ids format

tree_mat
Matrix format

label_probs
List

label_costs
List

pname_from_to
Dataframe name, from, to

cname_from_to
Dataframe name, from, to

hname_from_to

Value

Long format dataframe for input to `define_model()` or `dectree()`. Note this is for a single edge value type (e.g. cost or QALY). Use `run_cedectree()` for cost and health.

See also