Insert to transition matrix

Insert values into a transition matrix from a long array of from, to values

Insert values into a cost matrix from a long array of from, to values

Insert values into a probability matrix from a long array of from, to values

Usage

insert_to_transmat(dat, mat)

insert_to_costmat(dat, mat)

insert_to_probmat(dat, mat)

Arguments

dat

Long from-to-vals array. Either cost or prob.

mat

Transition matrix

Value

matrix

transition matrix

Probability matrix

Examples

#insert_to_transmat(dat, mat)