Branch Joint Probabilities

Provides a measure of the chances of following particular paths.

branch_joint_probs(model, ...)

Arguments

model
Branch conditional probabilities (matrix)

nodes
Which nodes to return; default to all

Value

Transition matrix with joint probabilities

Details

These probabilities could be used to weight branch costs or QALYs to indicate the relative contribution to the total expected value.

Examples

model <-
  define_model(
    transmat =
      list(prob =
             matrix(data = c(NA, 0.5, 0.5, NA,  NA,  NA,  NA,
                             NA, NA, NA,   0.1, 0.9, NA,  NA,
                             NA, NA, NA,   NA,  NA,  0.9, 0.1),
                    nrow = 3,
                    byrow = TRUE),
           vals =
             matrix(data = c(NA, 1,  5,  NA, NA, NA, NA,
                             NA, NA, NA, 1,  9,  NA, NA,
                             NA, NA, NA, NA, NA, 9,  1),
                    nrow = 3,
                    byrow = TRUE)))
model

#> $prob
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,]   NA  0.5  0.5   NA   NA   NA   NA
#> [2,]   NA   NA   NA  0.1  0.9   NA   NA
#> [3,]   NA   NA   NA   NA   NA  0.9  0.1
#> 
#> $vals
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7]
#> [1,]   NA    1    5   NA   NA   NA   NA
#> [2,]   NA   NA   NA    1    9   NA   NA
#> [3,]   NA   NA   NA   NA   NA    9    1
#> 
#> attr(,"class")
#> [1] "transmat" "list"    

branch_joint_probs(model)

#> Error in UseMethod("branch_joint_probs", model): no applicable method for 'branch_joint_probs' applied to an object of class "c('transmat', 'list')"

branch_joint_probs(model)*model$vals

#> Error in UseMethod("branch_joint_probs", model): no applicable method for 'branch_joint_probs' applied to an object of class "c('transmat', 'list')"

df <-
  data.frame(
    from = c(1,2,1),
    to = c(2,3,4),
    prob = c(0.1,0.5,0.9),
    vals = c(1,2,3))

branch_joint_probs.dat_long(df, 4)

#> Error in branch_joint_probs.dat_long(df, 4): could not find function "branch_joint_probs.dat_long"

#0.9

branch_joint_probs.dat_long(df, 3)

#> Error in branch_joint_probs.dat_long(df, 3): could not find function "branch_joint_probs.dat_long"

#0.1*0.5

mod <- define_model(dat_long = df)
branch_joint_probs(mod, 3)

#> Error in UseMethod("branch_joint_probs", model): no applicable method for 'branch_joint_probs' applied to an object of class "c('dat_long', 'data.frame')"

branch_joint_probs(mod, 3)[[1]] %>% cumprod()

#> Error in UseMethod("branch_joint_probs", model): no applicable method for 'branch_joint_probs' applied to an object of class "c('dat_long', 'data.frame')"